refactored the gui code.
